+++ src/sagemaker/tensorflow/serving.py
@@ -123,7 +123,7 @@ class Model(sagemaker.Model):
         self._framework_version = framework_version
         self._container_log_level = container_log_level
 
-    def prepare_container_def(self, instance_type):
+    def prepare_container_def(self, instance_type, accelerator_type=None):
         image = self._get_image_uri(instance_type)
         env = self._get_container_env()
         return sagemaker.container_def(image, self.model_data, env)