+++ tests/conftest.py
@@ -22,7 +22,9 @@ from sagemaker import Session
 from sagemaker.local import LocalSession
 from sagemaker.chainer import Chainer
 from sagemaker.mxnet import MXNet
+from sagemaker.rl import RLEstimator
 from sagemaker.pytorch.defaults import PYTORCH_VERSION
+from sagemaker.sklearn.defaults import SKLEARN_VERSION
 from sagemaker.tensorflow.defaults import TF_VERSION
 
 
@@ -33,10 +35,13 @@ def pytest_addoption(parser):
     parser.addoption('--sagemaker-client-config', action='store', default=None)
     parser.addoption('--sagemaker-runtime-config', action='store', default=None)
     parser.addoption('--boto-config', action='store', default=None)
-    parser.addoption('--tf-full-version', action='store', default=TF_VERSION)
-    parser.addoption('--mxnet-full-version', action='store', default=MXNet.LATEST_VERSION)
     parser.addoption('--chainer-full-version', action='store', default=Chainer.LATEST_VERSION)
+    parser.addoption('--mxnet-full-version', action='store', default=MXNet.LATEST_VERSION)
     parser.addoption('--pytorch-full-version', action='store', default=PYTORCH_VERSION)
+    parser.addoption('--rl-coach-full-version', action='store', default=RLEstimator.COACH_LATEST_VERSION)
+    parser.addoption('--rl-ray-full-version', action='store', default=RLEstimator.RAY_LATEST_VERSION)
+    parser.addoption('--sklearn-full-version', action='store', default=SKLEARN_VERSION)
+    parser.addoption('--tf-full-version', action='store', default=TF_VERSION)
 
 
 @pytest.fixture(scope='session')
@@ -79,10 +84,8 @@ def sagemaker_local_session(boto_config):
     return LocalSession(boto_session=boto_session)
 
 
-@pytest.fixture(scope='module', params=['1.4', '1.4.1', '1.5', '1.5.0', '1.6', '1.6.0',
-                                        '1.7', '1.7.0', '1.8', '1.8.0', '1.9', '1.9.0',
-                                        '1.10', '1.10.0', '1.11', '1.11.0'])
-def tf_version(request):
+@pytest.fixture(scope='module', params=['4.0', '4.0.0', '4.1', '4.1.0', '5.0', '5.0.0'])
+def chainer_version(request):
     return request.param
 
 
@@ -92,19 +95,41 @@ def mxnet_version(request):
     return request.param
 
 
-@pytest.fixture(scope='module', params=["0.4", "0.4.0"])
+@pytest.fixture(scope='module', params=['0.4', '0.4.0'])
 def pytorch_version(request):
     return request.param
 
 
-@pytest.fixture(scope='module', params=['4.0', '4.0.0', '4.1', '4.1.0', '5.0', '5.0.0'])
-def chainer_version(request):
+@pytest.fixture(scope='module', params=['0.20.0'])
+def sklearn_version(request):
+    return request.param
+
+
+@pytest.fixture(scope='module', params=['1.4', '1.4.1', '1.5', '1.5.0', '1.6', '1.6.0',
+                                        '1.7', '1.7.0', '1.8', '1.8.0', '1.9', '1.9.0',
+                                        '1.10', '1.10.0', '1.11', '1.11.0'])
+def tf_version(request):
+    return request.param
+
+
+@pytest.fixture(scope='module', params=['0.10.1', '0.10.1', '0.11', '0.11.0'])
+def rl_coach_tf_version(request):
+    return request.param
+
+
+@pytest.fixture(scope='module', params=['0.11', '0.11.0'])
+def rl_coach_mxnet_version(request):
+    return request.param
+
+
+@pytest.fixture(scope='module', params=['0.5', '0.5.3'])
+def rl_ray_version(request):
     return request.param
 
 
 @pytest.fixture(scope='module')
-def tf_full_version(request):
-    return request.config.getoption('--tf-full-version')
+def chainer_full_version(request):
+    return request.config.getoption('--chainer-full-version')
 
 
 @pytest.fixture(scope='module')
@@ -118,5 +143,20 @@ def pytorch_full_version(request):
 
 
 @pytest.fixture(scope='module')
-def chainer_full_version(request):
-    return request.config.getoption('--chainer-full-version')
+def rl_coach_full_version(request):
+    return request.config.getoption('--rl-coach-full-version')
+
+
+@pytest.fixture(scope='module')
+def rl_ray_full_version(request):
+    return request.config.getoption('--rl-ray-full-version')
+
+
+@pytest.fixture(scope='module')
+def sklearn_full_version(request):
+    return request.config.getoption('--sklearn-full-version')
+
+
+@pytest.fixture(scope='module')
+def tf_full_version(request):
+    return request.config.getoption('--tf-full-version')